Add function to disconnect tunnelblick
Prior to this commit, when the network adapter would switch between
ethernet and wifi if tunnelblick had an open vpn connection then
nothing work work because it is still somehow bounde to the
adapter that was switched off.

After this commit, we simply disconnect vpn connections whenever
we swap between adapters.